Add files via upload
This project aims to provide movie recommendations based on user input. Using a dataset with 4000+ movies, I've implemented a recommendation system that calculates the similarity between the user's input movie name and the movies in the dataset using the 'cosine_similarity' function of sklearn library of python. This approach allows to suggest movies that are similar in content or style to the one provided by the user, enhancing their movie-watching experience.